# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: DunningDog API
  version: 1.0.0
  description: |
    Stable v1 API contracts for DunningDog.

    ## Authentication

    The API supports two authentication methods:

    - **Bearer token** (JWT) — for browser sessions via `Authorization: Bearer <token>`
    - **API key** — for programmatic access via `x-api-key: dd_live_...` header (Scale plan only)

    API keys are scoped. Each endpoint requires specific scopes listed below.

    ## Scopes

    | Scope | Description |
    |-------|-------------|
    | `read:dashboard` | Read dashboard summary and metrics |
    | `read:recoveries` | Read and export recovery attempts |
    | `read:sequences` | Read dunning sequences |
    | `write:sequences` | Create and update dunning sequences |
    | `read:settings` | Read workspace settings (branding, notifications) |
  license:
    name: Proprietary
    url: https://dunningdog.com/legal/api-license
...
servers:
  - url: https://dunningdog.com
    description: Production
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /api/dashboard/export:
    get:
      summary: Export recovery attempts as CSV
      operationId: exportRecoveries
      x-required-scope: read:recoveries
      parameters:
        - name: startDate
          in: query
          required: false
          description: ISO 8601 date. Defaults to 30 days before endDate.
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: false
          description: ISO 8601 date. Defaults to today.
          schema:
            type: string
            format: date
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/RecoveryStatus"
      responses:
        "200":
          description: CSV file download
          content:
            text/csv:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/UnauthorizedProblem"
        "403":
          $ref: "#/components/responses/ForbiddenProblem"
        default:
          $ref: "#/components/responses/ProblemResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for programmatic access (Scale plan only). Create keys in Settings > API Keys.
  responses:
    ProblemResponse:
      description: Error response in Problem+JSON format
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    UnauthorizedProblem:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ForbiddenProblem:
      description: Forbidden
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, code, traceId]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        traceId:
          type: string
        meta:
          type: object
          additionalProperties: true
...
    RecoveryStatus:
      type: string
      enum: [pending, recovered, failed, abandoned]
...